Convert component to Typescript
